Cause the X server to terminate if the WM internal client thread exits due to an error
We can't call FatalError from an internal client thread, so these failures are
silent at the moment (in the sense that no pop-up occurs to tell the user that
we are terminating or why).

Taking the server down is kind of harsh, perhaps we should try to restart the WM
internal client thread a few times, after an increasing interval.

Also, I'm just noticing I should be using pthread_kill() to ensure the signal is
delivered to the thread which is running the main dispatch loop so that will
exit the select() it may be blocked in and notice that DE_TERMINATE is set.

v2: Use raise() rather than kill() for MinGW portability

Use C89 raise(SIGTERM) rather than the POSIX equivalent kill(getpid(), SIGTERM).

This still isn't quite right: On Cygwin for a clean shutdown we need send the
signal to the server thread so it interrupts select(), but it's unclear that can
happen at all for MinGW.  Restarting the WM thread seems like a better
alternative.

Signed-off-by: Jon Turney <jon.turney@dronecode.org.uk>