Revert "glx: Simplify glXDestroyContext"
This reverts commit 7f5adf73a0f9a951a6df201532b4031d38054369.

This seems to miss the whole point of the idExists flag, as it makes the
lifetime of that being true the same as the lifetime of the Context resource.

This leasds to BadContextTag errors when the tag for a deleted context is given
as the previous tag in a MakeContextCurrent request.