Fix Xephyr initial window sizing when run without a window manager
Conversion to xcb dropped a XResizeWindow() from hostx_scren_init(), so now
Xephyr assumes that it will receive a XCB_CONFIGURE_NOTIFY with the window's
initial placement after the window is mapped.

This assumption is not true when running without a WM, or when the WM is
deficient.

Restore setting the window size before it is mapped to ensure the window gets
sized correctly.

See fd.o bug #74849 https://bugs.freedesktop.org/show_bug.cgi?id=74849

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>