Use raise() rather than kill()
MinGW portability fix for "Cause the X server to terminate if the WM internal
client thread exits due to an error".

Use C89 raise(SIGTERM) rather than the POSIX equivalent kill(getpid(), SIGTERM).

This still isn't quite right: On Cygwin for a clean shutdown we need send the
signal to the server thread so it interrupts select(), but it's unclear that can
happen at all for MinGW.  Restarting the WM thread seems like a better
alternative.

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>