Fix appearance of windows with only WM_NORMAL_HINTS
Now that XGetWMNormalHints() is actually being used, we can have a window with
default style, but HINT_NOMAXIMIZE, but this doesn't get given the default
style. (Testcase: Xephyr)

This is just a quick workaround.  To fix this properly, we should probably
rewrite winApplyHints() so it starts from a default and then modifies it
depending on the hints, rather than trying to set the default style at the end
if we guess no other style has been set.

Signed-off-by: Jon Turney <jon.turney@dronecode.org.uk>