Fix sometimes missing sysmenu icon on Vista and later
It seems that the way we create windows makes Vista display them initially
without a sysmenu icon.  If the application subsequently changes it's icon, all
works well, but if the icon is set before the window is displayed and never
changes, the window is shown without a sysmenu icon.

Rearrange things to avoid this behaviour:

Explicitly select default icons instead of assuming class icons will be used for
a window created without the WS_SYSMENU style.

It also seems we must send WM_SETICON to update the icon *after* we have updated
style to include WS_SYSMENU in order for the icon to actually be used.

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>