Add Windows-DRI extension
If windowsdriproto headers are available, build a Windows-DRI extension, which
supports requests to enable local clients to directly render to a Windows
drawable:

- a query to check if WGL is being used on a screen
- a query to map a fbconfigID to a native pixelformatindex
- a query to map a drawable to a native handle

Windows-DRI can only be useful if we are using WGL, so make an note if WGL is
active on a screen

Add glxWindDeferredCreateDrawable(), and move deferred drawable creation code
there

Modify glxWinDeferredCreateDrawable(), so that pixmaps are placed into a file
mapping, so they exists in memory which can be shared with the direct rendering
process

XXX: Currently, this file mapping is accessed by a name generated from the XID.
This will not be unique across multiple server instances. It would perhaps be
better, although more complicated, to use an anonymous file mapping, and then
duplicate the handle for the direct rendering process.

Adjust glxWinSetPixelFormat() so it doesn't require a context, just a config

v2:
Various printf format warning fixes