Improve handling of some error conditions associated with xkbcomp temporary file
Fix a compilation conditional so that a failure to write to the xkbcomp
temporary file is correctly reported, and reports the problematic filename.

Improve Win32TempDir() so that the directory given by the TEMP and TMP
environment variables is checked for accessibility, so that specifying a
non-existent directory doesn't cause a failure.

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>