Remove duplicate LogInit() calls
Remove LogInit() during ddxGiveUp()

Remove g_fLogInited

Remove ddxUseMsg() not reporting the actual problem with the command line, and
saying we must read logfile for options

ddxProcessArgument() needs further auditing:
- For unrecognized options, instead of calling UseMsg(), return 0 which means
that ProcessCommandLine() will report the option as unrecognized and call
UseMsg().
- Where we should provide a real diagnostic, call FatalError()

This probably also needs more attention for MinGW: I guess because XWin is
marked as a GUI application, there is no console for help output to go to.
Possibly that can be fixed with an AttachConsole(), rather than trying to
arrange things so it goes to the logfile...

Signed-off-by: Jon Turney <jon.turney@dronecode.org.uk>