Add crash reporting
xorg_backtrace() is called in a few places to generate a backtrace for the log,
when the condition is not fatal (mi queue overflow, vpnprintf unrecognized
format), where we don't want to run the crash reporter

So add a separate, explicit invocation of the crash reporter, which is invoked
on a fatal signal only

v2:
Rather than invoking an external script to do crashreporting, load an optional
DLL which contains the crashreporter.

Use new interfaces in Cygwin 2.0.0 to extract EXCEPTION_POINTERS from the signal
information, and pass those to the crashreporter.

v3:
Downgrade crashreporter not loaded message to debug

v4:
Fix building on non-Windows