Ensure that XRANDR reports an output
If using the X server internal XRANDR 1.0 interface, it seems we must register a
display size with RRRegisterSize() in order to make RRGetInfo() call
RRScanOldConfig() to generate an output and crtc for us.

Without this, the following GDM bug means that an XDMCP session to GDM cannot be
started.

https://bugzilla.gnome.org/show_bug.cgi?id=736054

Instead, use the more recent XRANDR 1.2 internal interface to explicitly create
the output and crtc, and maintain a single mode which represents the current
display size.

Also don't emit a RRScreenSizeNotify when a RRScreenSizeSize is done which has
no effect, this seems to throw the GDM greeter into a loop...

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>