Ensure that XRANDR reports an output
If using the X server internal XRANDR 1.0 interface, it seems we must register a
display size with RRRegisterSize() in order to make RRGetInfo() call
RRScanOldConfig() to generate an output and crtc for us.

Without this, the following GDM bug means that an XDMCP session to GDM cannot be
started.

https://bugzilla.gnome.org/show_bug.cgi?id=736054

Instead, use the more recent XRANDR 1.2 internal interface to explicitly create
the output and crtc, and maintain a single mode which represents the current
display size.

Also don't emit a RRScreenSizeNotify when a RRScreenSizeSize is done which has
no effect, this seems to throw the GDM greeter into a loop...

v2:
Maintain reference count for the mode we maintain more correctly, to avoid
double free causing a crash on shutdown

Connect crtc to output, so a subsequent RRSetCrtcConfig request doesn't change
anything, so we don't fail due to our lack of rrSetConfig or rrCrtcSet hooks.

See https://cygwin.com/ml/cygwin-xfree/2015-02/msg00032.html