Revert "os: -displayfd option fixes"
This reverts the remaining part of commit 16e7057efee415d5b6daacb8b819d714086b4a05.
It's not totally clear what that is supposed to be fixing.

Since 1.17 and '-nolisten tcp' by default, there is a problem when we try to
start an X server on a display number which is already in use: As Xtrans is
currently written, a failed attempt to open the UNIX domain socket ends with
TransSocketUNIXClose() deleting the socket.

This prevents any future attempts to connect to the socket.

Prior to 1.17, an attempt to bind to the TCP/IP socket for the display was made
and failed first, so this never heppened.

This makes sure that taking the lock file is attempted before opening sockets,
which avoids this problem in the non-'-displayfd' case.

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>