Update WM_STATE, _NET_WM_STATE when window state changes
This avoids a problem seen with evolution, which updates it WM_HINTS to change
icon when new mail it received.

If started maximized, this maximization is re-applied with every WM_HINTS
change, even if it has been minimized.

v2:
Fix broken Qt menus

Qt won't map unmapped windows which have an incorrect WM_STATE of NormalState.
So update WM_STATE when window is unmapped as well.