Drop GLXServerExtensions string constant listing GLX extensions for swrast
Since 2d719433, it looks like we are free()-ing this constant string when we
build the real extension list string, which is perhaps not ideal. Instead
initialize GLXextensions to strdup(""), as the other dri loaders do.

The GLX extension list for swrast is now generated dynamically, so the string
constant is no longer needed.

Signed-off-by: Jon Turney <jon.turney@dronecode.org.uk>