Re-order window creation process
Be a bit more careful to do things in the right order and set the style flags
and window state before we show the window.

This is probably the right thing to do in any case as it means we can avoid the
messy appearance of the window changing style just after it is first shown

This is achieved by having WM_CREATE send a WM_WM_CREATE message to our window
manager thread, which then does all the work of discovering the windows style.

Note that this subtly changes the semantics of winCreateWindowsWindow():
previously the window was visible and drawn before that function returned, now
that happens asychronously; I'm not sure if that could cause problems or not...