Fix a problem with initially maximized windows
Fix a problem which occurs when a window is initially maxmimized: It's shown
unmaximized, then maximized.  The initial unmaximized show causes a
WM_WM_CHANGE_STATE to unmaximized to be sent, while a SW_MAXIMIZE is in flight.
These overlapping in-flight messages can cause the window to get stuck flipping
between maximized and unmaximized states.

It seems that maximizing a Windows window is not a state, but an action which
can only be applied to a visible window.  So, this can't be fixed by maximizing
the window before it is made visible.

Instead, explicitly sequence things so the window is always initially shown
unmaximized, then start tracking the state, and then maximize it.