Add -icon option to set the screen window icon in windowed mode
Add an -icon option to set the screen window icon in windowed mode
Allow cygwin paths in an icon-specification
Update man pages and system.XWinrc appropriately

Also, log an error if the icon specified for TRAYICON cannot be loaded

Also, fix a bug in appending a '\' to IconDirectory only if it doesn't
already end with one, which was fortunately benign.

Also, if we are exiting in ddxProcessArgument() due to an error in options,
use a non-zero exit status.

XXX: LoadImageComma would be simpler if we just said that XWinrc paths are
Cygwin paths on Cygwin, Windows paths on MinGW, but that could break
existing .XWinrc files

XXX: Given that we can specify paths, I'm not sure what IconDirectory wins
us.

v2:
Fix formatting problems in man page additions