Unmap minimized windows in rootless modes
At the moment, when a window is minimized, it is moved to (0,0) on the shadow
framebuffer.  This can leads to various problems: tooltips and menus can 'punch
through' from an apparently minimized window if the mouse pointer is positioned
over the native frame of another X window which covers the sensitive regions of
the apparently minimized window.  I think may also be leading to some repainting
glitches when uncovering areas covered by native window frames.

So, don't move minimized windows to (0,0), simply unmap them.

This has the disadvantage that the contents of windows which are minimized are
no longer stored in the shadow frame buffer, so must be redrawn when restored,
but I assume they were being re-exposed when restored, anyhow...

Signed-off-by: Jon Turney <jon.turney@dronecode.org.uk>