Add logging which records if we are in an RDP session
In an RDP session, the screen bpp will be set to the limited depth allowed
for RDP sessions (16 bpp by default)

This seems to sometimes cause problems with an X server started when the
screen bpp was more, either because the bpp conversion introduces horrible
dithering artefacts, or causes DirectDraw failures which we don't handle
well

Perhaps we should force engine 1 when started in a RDP session? Kind of a
bodge and doesn't really solve the problem above.