Update X.org logo icon
Generate icon from X.org logo SVG, rather than using a hand-crafted
icon.

This SVG has been tweaked to put a thin white border around the X, so it is
visible on a black background.

v2:
Also generate from svg in meson.build
Install ImageMagick in CI for convert tool

XXX: compile_resources() lacks a depend: keyword to make it depend on the
logo icon conversion.  It's also not clear how the directory into which the
output of the custom target is generated should be obtained to give to
windres as an include directory....

See https://github.com/mesonbuild/meson/issues/2789