Add an option to use alpha channel in multiwindow mode
Add an option to turn on the use of the X windows alpha channel in
multiwindow mode.  This works on W7/Vista (using
DwmEnableBlurBehindWindow()), and Windows 10 (using the undocumented
SetWindowCompositionAttribute()), but not on Windows 8/8.1

-compositewm must be enabled for this to be useful, as the X window only
uses a pixmap with an alpha channel when the Composite extensions is
enabled.

Future work:

A window property to control use of alpha?
Option to turn off blur on W7/Vista
Implement _NET_WM_WINDOW_OPACITY