Always set the X input focus to none when an X window loses focus.
We need to cancel the X input focus if the native input focus has gone to
the desktop, or another application's window.

We could maybe avoid some unneeded work by not doing this if the
WM_KILLFOCUS wParam is another window owned by us, which is immediately
going to be given the X input focus.