Only convert WM_DESTROY into a WM_DELETE_WINDOW for unowned windows
winDestroyWindowsWindow() traverses the tree of child Windows windows,
marking them as destroyed, but in the Windows world there is another
class of windows which will get destroyed when another windows is: owned
windows.

We make top-level TRANSIENT_FOR windows owned (so they are minimized at
the same time at their owner). There's no easy way to enumerate the
owned windows of a Windows window.

Instead, we avoid translating a WM_DESTROY into a WM_DELETE_WINDOW
unless the window is unowned.  I'm uncertain under what circumstances
the kill here actually gets used.

See https://cygwin.com/ml/cygwin/2019-06/msg00168.html

In that case, we have a submenu which is TRANSIENT_FOR a menu, which is
TRANSIENT_FOR the applications main window.  When focus moves away, the
menu is unmapped, destroying it's window, which causes the submenu to be
also destroyed.