hw/xwin: Only load the Windows-DRI extension if WGL is enabled
Only load the Windows-DRI extension if WGL is enabled
Also do a bit of de-cruftification on loading that extension

This avoids an unfortunate interaction in the client: libGL loads
opengl32 to determine if the renderer is blacklisted after checking if
the display has the Windows-DRI extension, but before it checks if
Windows-DRI is available on a screen.

Normally, this is no problem.  But some ATI drivers like to terminate
your application with a "LoadLibrary failed with error 87: The parameter
is incorrect" MessageBox, if you try to load opengl32 under RDP.

This still isn't perfect as we might not be able to use WGL (which we
don't find out until the screen is initialized), even if it is enabled,
but should mean you can at least get a working server by using -nowgl in
that problematic situation.