Avoid mis-rendering in some multi-monitor configurations
Avoid mis-rendering in multi-monitor configurations with disparate DPI
values, due to the non-client area scaling enabled with dpiAwareness set
to PerMonitorV2.

If you turn off 'Hide Root Window' from the tray menu, you can kind of
see what's going wrong: normally the client area of the X window is
exactly aligned with that in the root window, but when the non-client
area is rescaled by a DPI change (the title bar changes size
significantly), it's misaligned so that part of the X window is outside
the client area (and thus clipped in updates).

I think this is due to the AdjustWindowsRectEx() function not being DPI
aware (I guess it's always computing the non-client window rect based on
the processes initial DPI)

Unfortunately, from an initial look, rewriting things to use
AdjustWindowRectExForDpi() isn't trivial (since we need to make 'DPI of
the monitor this Window is going to end up on' available to it)

So for the moment, I think I'll apply your reversion (although this
probably comes at the cost of not scaling the window frame, the traymenu
and About... dialog)

See https://cygwin.com/pipermail/cygwin/2022-January/250544.html et seq.